+++ iec_api/usage_calculator/rates.py
@@ -1,5 +1,6 @@
 from dataclasses import dataclass, field
 from datetime import datetime
+from decimal import Decimal
 
 from mashumaro import DataClassDictMixin, field_options
 
@@ -9,6 +10,6 @@ class Rates(DataClassDictMixin):
     """Calculator Rates dataclass."""
 
     last_updated: datetime = field(metadata=field_options(alias="lastUpdated"))
-    home_rate: float = field(metadata=field_options(alias="homeRate"))
-    general_rate: float = field(metadata=field_options(alias="generalRate"))
-    vat: float = field(metadata=field_options(alias="vat"))
+    home_rate: Decimal = field(metadata=field_options(alias="homeRate"))
+    general_rate: Decimal = field(metadata=field_options(alias="generalRate"))
+    vat: Decimal = field(metadata=field_options(alias="vat"))
